+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/entities/SurveyStatsReportUser.java
@@ -70,7 +70,7 @@ public class SurveyStatsReportUser
     private long partiallyCompleted;
     
     @Column( name = "complete_percentage")
-    private long completePercentage;
+    private double completePercentage;
     
     @Column( name = "delta")
     private long delta;
@@ -339,14 +339,14 @@ public class SurveyStatsReportUser
 
 
 
-    public long getCompletePercentage()
+    public double getCompletePercentage()
     {
         return completePercentage;
     }
 
 
 
-    public void setCompletePercentage( long completePercentage )
+    public void setCompletePercentage( double completePercentage )
     {
         this.completePercentage = completePercentage;
     }
